Updated test runner template to use JsonReporter. Changed parsing of JSON to strict=False and used Element.html instead of Element.text to deal with newline issues.
Changes to the SuiteRunner, SuiteRunnerFactory, and Browser
to ensure that all the pieces communicate consistently.

Terminate browser sessions when tool finishes (even if error)

Exit code indicates success/failure of the entire test suite.

Lettuce test modifications and improvements.  The tests are still
not passing, but most of the required changes are to the data the tests use.